src/requirements.txt
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.1
requests==2.26.0
MarkupSafe==2.0.1
django-extensions==3.1.3
PyJWT==2.1.0
celery==5.1.2
redis==3.5.3
flower==1.0.0
django-celery-beat==2.2.1
xlsxwriter==3.0.1
pandas==1.3.2
sentry-sdk==1.3.1
drf-yasg==1.20.0
Pillow==8.3.1
qrcode==7.3
pycryptodome==3.10.1
zstd==1.5.0.2
